feat : 지난 공연에 대해 선물하기 미노출
